Update for travel rule soft-match/manual review
Be able to specify that we are in manual review state. Then allow manual review state to specify that we have a soft match and other side can provide free-form KYC data.

This also results in changing the KYC data object status field to no longer be a pure enum but instead have the ability to describe why we aborted. This is also probably useful for other cases where we abort