eCommerce P2M Payments: DIP-158  (#166)
* Update dip-8.md

* Update lip-8.md:

Disclaimer regarding the discovery phase
Extended scope for fund pull pre-approval
Revoke an authorization in more details

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lip-8.md

* Apply suggestions from code review

Co-authored-by: kphfb <kph@fb.com>

* PR fixes

* split DIP-8 to funds pre approval and auth/capture

* add dip 158 for p2m charge + auth/capture

* Created DIP-158 (AKA P2M DIP).

* Incorporated additional comments

* Update dip-8.md

* Update lip-8.md:

Disclaimer regarding the discovery phase
Extended scope for fund pull pre-approval
Revoke an authorization in more details

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lips/lip-8.md

Co-authored-by: kphfb <kph@fb.com>

* Update lip-8.md

* Apply suggestions from code review

Co-authored-by: kphfb <kph@fb.com>

* PR fixes

* split DIP-8 to funds pre approval and auth/capture

* add dip 158 for p2m charge + auth/capture

* Created DIP-158 (AKA P2M DIP).

* Incorporated additional comments

* Revert DIP-8 initial changes draft

* Update dips/dip-158.md

Co-authored-by: dahliamalkhi <dahliamalkhi@outlook.com>

* Update dips/dip-158.md

Co-authored-by: dahliamalkhi <dahliamalkhi@outlook.com>

* Update dips/dip-158.md

Co-authored-by: dahliamalkhi <dahliamalkhi@outlook.com>

* Update dips/dip-158.md

Co-authored-by: dahliamalkhi <dahliamalkhi@outlook.com>

* Update dips/dip-158.md

Co-authored-by: dahliamalkhi <dahliamalkhi@outlook.com>

* Updated based on latest comments from Dahlia

* Added glossary item for Reference ID based on Dahlia's comments.

* Update dips/dip-158.md

Co-authored-by: David Wolinsky <isaac.wolinsky@gmail.com>

* Update dips/dip-158.md

Co-authored-by: David Wolinsky <isaac.wolinsky@gmail.com>

* Updated based on David W. comments

* Additional comments following meeting with the team.
Notes:
Agreed that we will create new structures for P2M purposes rather than impose changes on existing structures from 0 < DIP < 2  (did not want to use the actual DIP name :)).
For example, we will NOT extend or change the PaymentActorObject. Instead, we will create a new object suitable for P2M needs (e.g. BusinessActorObject).
Agreed that business_data (i.e. merchant basic details) should be part of the InfoCommand response
Agreed that payer_data (i.e. payer information) should be part of the InitCommand(s) request
Agreed that, for the time being, there is no support for an additional step to ask for more data.
The receiving VASP can either accept or reject the payment based on the information provided.
We mentioned that the receiving VASP should be able to state "extra" information that is required on top of the payer_data (e.g. DOB might be required for buying alcohol) as part of the InfoCommand response.
This will be achieved by an additional optional element (type will be JSON)
A similar optional element will be added to the InitCommand(s) request. This element is to be used by the sending VASP to send the additional information requested (e.g. DOB)
Agreed (with a heavy heart...) to stop using the functionality names and stick with the command types
Agreed to use two different init command types (for each payment type) to replace the single InitCommand
We will change the name of the ReadyCommand to be ReadyNotification
Add dual attestation specification for transactions over the limit

* Revert DIP-8 to master version to remove from pull request

* Another cycle of comments incorporated.
Main changes:
* Removed DIP-8 fro the pull request
* Change InfoCommand name
* Added image_url to business_data
* Spelling and grammar issues

* Changed the order of sections and added internal links

* Internal links fix

* Appendix A link

* Added AddressObject and NationalIdObject

* * Introduced PaymentCommandErrorObject
* Removed usage of statuses

* add dual attestation info to charge command response and capture command request

* * Grammar issue
* Naming consistency
* Optional redirect URL for QR flow (POS scenario)

* Added sequence diagrams and partial capture support

* Update dip-158.md

Added diagram images

Co-authored-by: kphfb <kph@fb.com>
Co-authored-by: Karpick-E <59393941+Karpick-E@users.noreply.github.com>
Co-authored-by: Daniel Prinz <daniel@firstdag.com>
Co-authored-by: dahliamalkhi <dahliamalkhi@outlook.com>
Co-authored-by: David Wolinsky <isaac.wolinsky@gmail.com>